+++ FakeServer/Common/QueryHelper.cs
@@ -47,6 +47,19 @@ namespace FakeServer.Common
     {
         public static PaginationHeader GetPaginationHeader(string url, int totalCount, int skip, int take, string skipWord, string takeWord)
         {
+            if (skipWord == "page" && takeWord == "per_page")
+            {
+                var totalPages = totalCount / take + (totalCount % take != 0 ? 1 : 0);
+                var page = (skip / take) + 1;
+                return new PaginationHeader
+                {
+                    Prev = page > 1 ? $"{url}?{skipWord}={(page - 1)}&{takeWord}={take}" : string.Empty,
+                    Next = page < totalPages ? $"{url}?{skipWord}={(page + 1)}&{takeWord}={take}" : string.Empty,
+                    First = page != 1 ? $"{url}?{skipWord}=1&{takeWord}={take}" : string.Empty,
+                    Last = page != totalPages ? $"{url}?{skipWord}={totalPages}&{takeWord}={take}" : string.Empty
+                };
+            }
+
             return new PaginationHeader
             {
                 Prev = skip > 0 ? $"{url}?{skipWord}={(skip - take > 0 ? skip - take : 0)}&{takeWord}={(take - skip < 0 ? take : skip)}" : string.Empty,
@@ -79,7 +92,7 @@ namespace FakeServer.Common
                 ["results"] = results,
                 ["link"] = pg,
                 ["count"] = totalCount,
-                [options.SkipWord] = options.Skip,
+                [options.SkipWord] = options.SkipWord == "page" ? (options.Skip / options.Take) + 1 : options.Skip,
                 [options.TakeWord] = options.Take
             };
 
@@ -110,6 +123,21 @@ namespace FakeServer.Common
                 takeWord = "limit";
             }
 
+            if (queryParams.Contains("per_page"))
+            {
+                take = Convert.ToInt32(query["per_page"]);
+                queryParams.Remove("per_page");
+                takeWord = "per_page";
+            }
+
+            if (queryParams.Contains("page"))
+            {
+                var page = Convert.ToInt32(query["page"]);
+                skip = (page - 1) * take;
+                queryParams.Remove("page");
+                skipWord = "page";
+            }
+
             if (queryParams.Contains("q"))
             {
                 isTextSearch = true;
@@ -133,6 +161,22 @@ namespace FakeServer.Common
 
             return new QueryOptions { Skip = skip, Take = take, SkipWord = skipWord, TakeWord = takeWord, IsTextSearch = isTextSearch, 
                 Fields = fields, SortFields = sortFields, QueryParams = queryParams };
+        }
+
+        public static bool IsQueryValid(IQueryCollection query)
+        {
+            List<List<string>> validQuerySets = new List<List<string>>{
+                new List<string> { "skip", "take" },
+                new List<string> { "offset", "limit" },
+                new List<string> { "page", "per_page" }
+            };
+
+            var usedQuerySet = validQuerySets.FirstOrDefault(q => q.Any(query.ContainsKey));
+            if (usedQuerySet == null)
+                return true;
+
+            var notValidKeys = validQuerySets.Where(q => q != usedQuerySet).SelectMany(q => q);
+            return notValidKeys.Any(query.ContainsKey) == false;
         }
     }
 }